Move more content and remove cyclic dependencies
