+++ wfdb/__init__.py
@@ -17,16 +17,12 @@ from wfdb.io.annotation import (
     wrann,
     show_ann_labels,
     show_ann_classes,
-    ann2rr,
-    rr2ann,
-    csv2ann,
-    rdedfann,
     mrgann,
 )
 from wfdb.io.download import (
+    dl_files,
     get_dbs,
     get_record_list,
-    dl_files,
     set_db_index_url,
 )
 from wfdb.plot.plot import plot_items, plot_wfdb, plot_all_records
