+++ wfdb/io/_header.py
@@ -1,7 +1,6 @@
 import datetime
 import os
 import re
-import pdb
 
 import numpy as np
 import pandas as pd
