+++ wfdb/io/annotation.py
@@ -4,13 +4,12 @@ import os
 import pandas as pd
 import re
 import posixpath
-import struct
 import sys
 
 from wfdb.io import download
 from wfdb.io import _header
 from wfdb.io import record
-from wfdb.io.convert.edf import read_edf
+
 
 class Annotation(object):
     """
@@ -1928,7 +1927,7 @@ def rdann(
     if (pn_dir is not None) and ("." not in pn_dir):
         dir_list = pn_dir.split("/")
         pn_dir = posixpath.join(
-            dir_list[0], record.get_version(dir_list[0]), *dir_list[1:]
+            dir_list[0], download.get_version(dir_list[0]), *dir_list[1:]
         )
 
     return_label_elements = check_read_inputs(
@@ -2561,578 +2560,6 @@ def rm_last(*args):
     return
 
 
-def ann2rr(
-    record_name,
-    extension,
-    pn_dir=None,
-    start_time=None,
-    stop_time=None,
-    format=None,
-    as_array=True,
-):
-    from wfdb.processing import hr
-
-    """
-    Obtain RR interval series from ECG annotation files.
-
-    Parameters
-    ----------
-    record_name : str
-        The record name of the WFDB annotation file. ie. for file '100.atr',
-        record_name='100'.
-    extension : str
-        The annotatator extension of the annotation file. ie. for  file
-        '100.atr', extension='atr'.
-    pn_dir : str, optional
-        Option used to stream data from Physionet. The PhysioNet database
-        directory from which to find the required annotation file. eg. For
-        record '100' in 'http://physionet.org/content/mitdb': pn_dir='mitdb'.
-    start_time : float, optional
-        The time to start the intervals in seconds.
-    stop_time : float, optional
-        The time to stop the intervals in seconds.
-    format : str, optional
-        Print intervals in the specified format. By default, intervals are
-        printed in units of sample intervals. Other formats include
-        's' (seconds), 'm' (minutes), 'h' (hours). Set to 'None' for samples.
-    as_array : bool, optional
-        If True, return an an 'ndarray', else print the output.
-
-    Returns
-    -------
-    N/A
-
-    Examples
-    --------
-    >>> wfdb.ann2rr('sample-data/100', 'atr', as_array=False)
-    >>> 18
-    >>> 59
-    >>> ...
-    >>> 250
-    >>> 257
-
-    """
-    if (pn_dir is not None) and ("." not in pn_dir):
-        dir_list = pn_dir.split("/")
-        pn_dir = posixpath.join(
-            dir_list[0], record.get_version(dir_list[0]), *dir_list[1:]
-        )
-
-    ann = rdann(record_name, extension, pn_dir=pn_dir)
-
-    rr_interval = hr.calc_rr(ann.sample, fs=ann.fs)
-    rr_interval = np.insert(rr_interval, 0, ann.sample[0])
-
-    time_interval = rr_interval / ann.fs
-    if start_time is not None:
-        time_interval = time_interval[(time_interval > start_time).astype(bool)]
-    if stop_time is not None:
-        time_interval = time_interval[(time_interval < stop_time).astype(bool)]
-
-    # Already given in seconds (format == 's')
-    if format == "s":
-        out_interval = time_interval
-    elif format == "m":
-        out_interval = time_interval / 60
-    elif format == "h":
-        out_interval = time_interval / (60 * 60)
-    else:
-        out_interval = np.around(time_interval * ann.fs).astype(np.int)
-
-    if as_array:
-        return out_interval
-    else:
-        print(*out_interval, sep="\n")
-
-
-def rr2ann(rr_array, record_name, extension, fs=250, as_time=False):
-    """
-    Creates an annotation file from the standard input, which should usually
-    be a Numpy array of intervals in the format produced by `ann2rr`. (For
-    exceptions, see the `as_time` parameter below.). An optional second column
-    may be provided which gives the respective annotation mnemonic.
-
-    Parameters
-    ----------
-    rr_array : ndarray
-        A Numpy array consisting of the input RR intervals. If `as_time` is
-        set to True, then the input should consist of times of occurences. If,
-        the shape of the input array is '(n_annot,2)', then treat the second
-        column as the annotation mnemonic ('N', 'V', etc.). If a second column
-        is not specified, then the default annotation will the '"' which
-        specifies a comment.
-    record_name : str
-        The record name of the WFDB annotation file. ie. for file '100.atr',
-        record_name='100'.
-    extension : str
-        The annotatator extension of the annotation file. ie. for  file
-        '100.atr', extension='atr'.
-    fs : float, int, optional
-        Assume the specified sampling frequency. This option has no effect
-        unless the `as_time` parameter is set to convert to samples; in this
-        case, a sampling frequency of 250 Hz is assumed if this option is
-        omitted.
-    as_time : bool
-        Interpret the input as times of occurrence (if True), rather than as
-        samples (if False). There is not currently a way to input RR intervals
-        in time format between beats. For example, 0.2 seconds between beats
-        1->2, 0.3 seconds between beats 2->3, etc.
-
-    Returns
-    -------
-    N/A
-
-    Examples
-    --------
-    Using time of occurence as input:
-    >>> import numpy as np
-    >>> rr_array = np.array([[0.2, 0.6, 1.3], ['V', 'N', 'V']]).T
-    >>> wfdb.rr2ann(rr_array, 'test_ann', 'atr', fs=100, as_time=True)
-
-    Using samples as input:
-    >>> import numpy as np
-    >>> rr_array = np.array([4, 17, 18, 16])
-    >>> wfdb.rr2ann(rr_array, 'test_ann', 'atr')
-
-    """
-    try:
-        ann_sample = rr_array[:, 0]
-    except IndexError:
-        ann_sample = rr_array
-
-    if as_time:
-        ann_sample = (fs * ann_sample.astype(np.float64)).astype(np.int64)
-    else:
-        ann_sample = np.cumsum(ann_sample).astype(np.int64)
-
-    try:
-        ann_symbol = rr_array[:, 1].tolist()
-    except IndexError:
-        ann_symbol = rr_array.shape[0] * ['"']
-
-    wrann(record_name, extension, ann_sample, symbol=ann_symbol)
-
-
-def csv2ann(
-    file_name,
-    extension="atr",
-    fs=None,
-    record_only=False,
-    time_onset=True,
-    header=True,
-    delimiter=",",
-    verbose=False,
-):
-    """
-    Read a CSV/TSV/etc. file and return either an `Annotation` object with the
-    annotation descriptors as attributes or write an annotation file.
-
-    Parameters
-    ----------
-    file_name : str
-        The name of the CSV file to be read, including the '.csv' file
-        extension. If the argument contains any path delimiter characters, the
-        argument will be interpreted as PATH/BASE_RECORD. Both relative and
-        absolute paths are accepted. The BASE_RECORD file name will be used to
-        name the annotation file with the desired extension.
-    extension : str, optional
-        The string annotation file extension.
-    fs : float, optional
-        This will be used if annotation onsets are given in the format of time
-        (`time_onset` = True) instead of sample since onsets must be sample
-        numbers in order for `wrann` to work. This number can be expressed in
-        any format legal for a Python input of floating point numbers (thus
-        '360', '360.', '360.0', and '3.6e2' are all legal and equivalent). The
-        sampling frequency must be greater than 0; if it is missing, a value
-        of 250 is assumed.
-    record_only : bool, optional
-        Whether to only return the record information (True) or not (False).
-        If false, this function will generate the annotation file.
-    time_onset : bool, optional
-        Whether to assume the values provided in the 'onset' column are in
-        units of time (True) or samples (False). If True, convert the onset
-        times to samples by using the, now required, `fs` input.
-    header : bool, optional
-        Whether to assume the CSV has a first line header (True) or not
-        (False) which defines the signal names.
-    delimiter : str, optional
-        What to use as the delimiter for the file to separate data. The default
-        if a comma (','). Other common delimiters are tabs ('\t'), spaces (' '),
-        pipes ('|'), and colons (':').
-    verbose : bool, optional
-        Whether to print all the information read about the file (True) or
-        not (False).
-
-    Returns
-    -------
-    N/A : Annotation, optional
-        The WFDB Annotation object representing the contents of the CSV file
-        read.
-
-    Notes
-    -----
-    CSVs should be in one of the two possible following format:
-
-    1) All events are single time events (no duration).
-
-    onset,description
-    onset_1,description_1
-    onset_2,description_2
-    ...,...
-
-    Or this format if `header=False` is defined:
-
-    onset_1,description_1
-    onset_2,description_2
-    ...,...
-
-    2) A duration is specified for some events.
-
-    onset,duration,description
-    onset_1,duration_1,description_1
-    onset_2,duration_2,description_2
-    ...,...,...
-
-    Or this format if `header=False` is defined:
-
-    onset_1,duration_1,description_1
-    onset_2,duration_2,description_2
-    ...,...,...
-
-    By default, the 'onset' will be interpreted as a sample number if it is
-    strictly in integer format and as a time otherwise. By default, the
-    'duration' will be interpreted as time values and not elapsed samples. By
-    default, the 'description' will be interpreted as the `aux_note` for the
-    annotation and the `symbol` will automatically be set to " which defines a
-    comment. Future additions will allow the user to customize such
-    attributes.
-
-    Examples
-    --------
-    1) Write WFDB annotation file from CSV with time onsets:
-       ======= start example.csv =======
-       onset,description
-       0.2,p-wave
-       0.8,qrs
-       ======== end example.csv ========
-       >>> wfdb.csv2ann('example.csv', fs=360)
-       * Creates a WFDB annotation file called: 'example.atr'
-
-    2) Write WFDB annotation file from CSV with sample onsets:
-       ======= start example.csv =======
-       onset,description
-       5,p-wave
-       13,qrs
-       ======== end example.csv ========
-       >>> wfdb.csv2ann('example.csv', fs=10, time_onset=False)
-       * Creates a WFDB annotation file called: 'example.atr'
-       * 5,13 samples -> 0.5,1.3 seconds for onset
-
-    3) Write WFDB annotation file from CSV with time onsets, durations, and no
-       header:
-       ======= start example.csv =======
-       0.2,0.1,qrs
-       0.8,0.4,qrs
-       ======== end example.csv ========
-       >>> wfdb.csv2ann('example.csv', extension='qrs', fs=360, header=False)
-       * Creates a WFDB annotation file called: 'example.qrs'
-
-    """
-    # NOTE: No need to write input checks here since the Annotation class
-    # should handle them (except verifying the CSV input format which is for
-    # Pandas)
-    if header:
-        df_CSV = pd.read_csv(file_name, delimiter=delimiter)
-    else:
-        df_CSV = pd.read_csv(file_name, delimiter=delimiter, header=None)
-    if verbose:
-        print("Successfully read CSV")
-
-    if verbose:
-        print("Creating Pandas dataframe from CSV")
-    if df_CSV.shape[1] == 2:
-        if verbose:
-            print("onset,description format detected")
-        if not header:
-            df_CSV.columns = ["onset", "description"]
-        df_out = df_CSV
-    elif df_CSV.shape[1] == 3:
-        if verbose:
-            print("onset,duration,description format detected")
-            print("Converting durations to single time-point events")
-        if not header:
-            df_CSV.columns = ["onset", "duration", "description"]
-        df_out = _format_ann_from_df(df_CSV)
-    else:
-        raise Exception(
-            """The number of columns in the CSV was not
-                        recognized."""
-        )
-
-    # Remove extension from input file name
-    file_name = file_name.split(".")[0]
-    if time_onset:
-        if not fs:
-            raise Exception(
-                """`fs` must be provided if `time_onset` is True
-                            since it is required to convert time onsets to
-                            samples"""
-            )
-        sample = (df_out["onset"].to_numpy() * fs).astype(np.int64)
-    else:
-        sample = df_out["onset"].to_numpy()
-    # Assume each annotation is a comment
-    symbol = ['"'] * len(df_out.index)
-    subtype = np.array([22] * len(df_out.index))
-    # Assume each annotation belongs with the 1st channel
-    chan = np.array([0] * len(df_out.index))
-    num = np.array([0] * len(df_out.index))
-    aux_note = df_out["description"].tolist()
-
-    if verbose:
-        print("Finished CSV parsing... writing to Annotation object")
-
-    if record_only:
-        return Annotation(
-            record_name=file_name,
-            extension=extension,
-            sample=sample,
-            symbol=symbol,
-            subtype=subtype,
-            chan=chan,
-            num=num,
-            aux_note=aux_note,
-            fs=fs,
-        )
-        if verbose:
-            print("Finished creating Annotation object")
-    else:
-        wrann(
-            file_name,
-            extension,
-            sample=sample,
-            symbol=symbol,
-            subtype=subtype,
-            chan=chan,
-            num=num,
-            aux_note=aux_note,
-            fs=fs,
-        )
-        if verbose:
-            print("Finished writing Annotation file")
-
-
-def rdedfann(
-    record_name,
-    pn_dir=None,
-    delete_file=True,
-    info_only=True,
-    record_only=False,
-    verbose=False,
-):
-    """
-    This program returns the annotation information from an EDF+ file
-    containing annotations (with the signal name given as 'EDF Annotations').
-    The information that is returned if `info_only` is set to True is:
-        {
-            'onset_time': list of %H:%M:%S.fff strings denoting the annotation
-                          onset times,
-            'sample_num': list of integers denoting the annotation onset
-                          sample numbers,
-            'comment': list of comments (`aux_note`) for the annotations,
-            'duration': list of floats denoting the duration of the event
-        }
-    Else, this function will return either the WFDB Annotation format of the
-    information of the file if `record_only` is set to True, or nothing if
-    neither are set to True though a WFDB Annotation file will be created.
-
-    Parameters
-    ----------
-    record_name : str
-        The name of the input EDF record to be read.
-    pn_dir : str, optional
-        Option used to stream data from Physionet. The Physionet
-        database directory from which to find the required record files.
-        eg. For record '100' in 'http://physionet.org/content/mitdb'
-        pn_dir='mitdb'.
-    delete_file : bool, optional
-        Whether to delete the saved EDF file (False) or not (True)
-        after being imported.
-    info_only : bool, optional
-        Return, strictly, the information contained in the file as formatted
-        by the original WFDB package. Must not be True if `record_only` is
-        True.
-    record_only : bool, optional
-        Whether to only return the annotation information (True) or not
-        (False). If False, this function will generate a WFDB-formatted
-        annotation file. If True, it will return the object returned if that
-        file were read with `rdann`. Must not be True if `info_only` is True.
-    verbose : bool, optional
-        Whether to print all the information read about the file (True) or
-        not (False).
-
-    Returns
-    -------
-    N/A : dict, Annotation, optional
-        If 'info_only' is set to True, return all of the annotation
-        information needed to generate WFDB-formatted annotation files.
-        If 'record_only' is set to True, return the WFDB-formatted annotation
-        object generated by the `rdann` output. If none are set to True, write
-        the WFDB-formatted annotation file.
-
-    Notes
-    -----
-    The entire file is composed of (seen here:
-    https://www.edfplus.info/specs/edfplus.html#edfplusannotations):
-
-    HEADER RECORD (we suggest to also adopt the 12 simple additional EDF+ specs)
-    8 ascii : version of this data format (0)
-    80 ascii : local patient identification (mind item 3 of the additional EDF+ specs)
-    80 ascii : local recording identification (mind item 4 of the additional EDF+ specs)
-    8 ascii : startdate of recording (dd.mm.yy) (mind item 2 of the additional EDF+ specs)
-    8 ascii : starttime of recording (hh.mm.ss)
-    8 ascii : number of bytes in header record
-    44 ascii : reserved
-    8 ascii : number of data records (-1 if unknown, obey item 10 of the additional EDF+ specs)
-    8 ascii : duration of a data record, in seconds
-    4 ascii : number of signals (ns) in data record
-    ns * 16 ascii : ns * label (must be 'EDF Annotations')
-    ns * 80 ascii : ns * transducer type (must be whitespace)
-    ns * 8 ascii : ns * physical dimension (must be whitespace)
-    ns * 8 ascii : ns * physical minimum (e.g. -500 or 34, different than physical maximum)
-    ns * 8 ascii : ns * physical maximum (e.g. 500 or 40, different than physical minimum)
-    ns * 8 ascii : ns * digital minimum (must be -32768)
-    ns * 8 ascii : ns * digital maximum (must be 32767)
-    ns * 80 ascii : ns * prefiltering (must be whitespace)
-    ns * 8 ascii : ns * nr of samples in each data record
-    ns * 32 ascii : ns * reserved
-
-    ANNOTATION RECORD
-
-    Examples
-    --------
-    >>> ann_info = wfdb.rdedfann('sample-data/test_edfann.edf')
-
-    """
-    # Some preliminary checks
-    if info_only and record_only:
-        raise Exception(
-            "Both `info_only` and `record_only` are set. Only one "
-            "can be set at a time."
-        )
-
-    # According to the EDF+ docs:
-    #   "The coding is EDF compatible in the sense that old EDF software would
-    #    simply treat this 'EDF Annotations' signal as if it were a (strange-
-    #    looking) ordinary signal"
-    rec = read_edf(
-        record_name,
-        pn_dir=pn_dir,
-        delete_file=delete_file,
-        record_only=True,
-        rdedfann_flag=True,
-    )
-
-    # Convert from array of integers to ASCII strings
-    annotation_string = ""
-    for chunk in rec.p_signal.flatten().astype(np.int64):
-        if chunk + 1 == 0:
-            continue
-        else:
-            adjusted_hex = hex(
-                struct.unpack("<H", struct.pack(">H", chunk + 1))[0]
-            )
-            annotation_string += bytes.fromhex(adjusted_hex[2:]).decode("ascii")
-            # Remove all of the whitespace
-            for rep in ["\x00", "\x14", "\x15"]:
-                annotation_string = annotation_string.replace(rep, " ")
-
-    # Parse the resulting annotation string
-    onsets = []
-    onset_times = []
-    sample_nums = []
-    comments = []
-    durations = []
-    all_anns = annotation_string.split("+")
-    for ann in all_anns:
-        if ann == "":
-            continue
-        try:
-            ann_split = ann.strip().split(" ")
-            onset = float(ann_split[0])
-            hours, rem = divmod(onset, 3600)
-            minutes, seconds = divmod(rem, 60)
-            onset_time = f"{hours:02.0f}:{minutes:02.0f}:{seconds:06.3f}"
-            sample_num = int(onset * rec.sig_len)
-            duration = float(ann_split[1])
-            comment = " ".join(ann_split[2:])
-            if verbose:
-                print(
-                    f"{onset_time}\t{sample_num}\t{comment}\t\tduration: {duration}"
-                )
-            onsets.append(onset)
-            onset_times.append(onset_time)
-            sample_nums.append(sample_num)
-            comments.append(comment)
-            durations.append(duration)
-        except IndexError:
-            continue
-
-    if info_only:
-        return {
-            "onset_time": onset_times,
-            "sample_num": sample_nums,
-            "comment": comments,
-            "duration": durations,
-        }
-    else:
-        df_in = pd.DataFrame(
-            data={
-                "onset": onsets,
-                "duration": durations,
-                "description": comments,
-            }
-        )
-        df_out = _format_ann_from_df(df_in)
-        # Remove extension from input file name
-        record_name = record_name.split(os.sep)[-1].split(".")[0]
-        extension = "atr"
-        fs = rec.fs
-        sample = (df_out["onset"].to_numpy() * fs).astype(np.int64)
-        # Assume each annotation is a comment
-        symbol = ['"'] * len(df_out.index)
-        subtype = np.array([22] * len(df_out.index))
-        # Assume each annotation belongs with the 1st channel
-        chan = np.array([0] * len(df_out.index))
-        num = np.array([0] * len(df_out.index))
-        aux_note = df_out["description"].tolist()
-
-        if record_only:
-            return Annotation(
-                record_name=record_name,
-                extension=extension,
-                sample=sample,
-                symbol=symbol,
-                subtype=subtype,
-                chan=chan,
-                num=num,
-                aux_note=aux_note,
-                fs=fs,
-            )
-        else:
-            wrann(
-                record_name,
-                extension,
-                sample=sample,
-                symbol=symbol,
-                subtype=subtype,
-                chan=chan,
-                num=num,
-                aux_note=aux_note,
-                fs=fs,
-            )
-
-
 def mrgann(
     ann_file1,
     ann_file2,
@@ -3471,7 +2898,7 @@ def mrgann(
         )
 
 
-def _format_ann_from_df(df_in):
+def format_ann_from_df(df_in):
     """
     Parameters
     ----------
