+++ wfdb/io/convert/edf.py
@@ -6,8 +6,10 @@ import posixpath
 import struct
 
 import numpy as np
+import pandas as pd
 
-from wfdb.io.record import Record, rdrecord, SIG_UNITS, get_version
+from wfdb.io.annotation import Annotation, format_ann_from_df, wrann
+from wfdb.io.record import Record, rdrecord, SIG_UNITS
 from wfdb.io import _url
 from wfdb.io import download
 
@@ -124,7 +126,7 @@ def read_edf(
         if "." not in pn_dir:
             dir_list = pn_dir.split("/")
             pn_dir = posixpath.join(
-                dir_list[0], get_version(dir_list[0]), *dir_list[1:]
+                dir_list[0], download.get_version(dir_list[0]), *dir_list[1:]
             )
 
         file_url = posixpath.join(download.PN_INDEX_URL, pn_dir, record_name)
@@ -985,4 +987,215 @@ def wfdb_to_edf(
                 "WARNING: output contains an invalid character, {}, at byte {}".format(
                     val, i
                 )
+            )
+
+
+def rdedfann(
+    record_name,
+    pn_dir=None,
+    delete_file=True,
+    info_only=True,
+    record_only=False,
+    verbose=False,
+):
+    """
+    This program returns the annotation information from an EDF+ file
+    containing annotations (with the signal name given as 'EDF Annotations').
+    The information that is returned if `info_only` is set to True is:
+        {
+            'onset_time': list of %H:%M:%S.fff strings denoting the annotation
+                          onset times,
+            'sample_num': list of integers denoting the annotation onset
+                          sample numbers,
+            'comment': list of comments (`aux_note`) for the annotations,
+            'duration': list of floats denoting the duration of the event
+        }
+    Else, this function will return either the WFDB Annotation format of the
+    information of the file if `record_only` is set to True, or nothing if
+    neither are set to True though a WFDB Annotation file will be created.
+
+    Parameters
+    ----------
+    record_name : str
+        The name of the input EDF record to be read.
+    pn_dir : str, optional
+        Option used to stream data from Physionet. The Physionet
+        database directory from which to find the required record files.
+        eg. For record '100' in 'http://physionet.org/content/mitdb'
+        pn_dir='mitdb'.
+    delete_file : bool, optional
+        Whether to delete the saved EDF file (False) or not (True)
+        after being imported.
+    info_only : bool, optional
+        Return, strictly, the information contained in the file as formatted
+        by the original WFDB package. Must not be True if `record_only` is
+        True.
+    record_only : bool, optional
+        Whether to only return the annotation information (True) or not
+        (False). If False, this function will generate a WFDB-formatted
+        annotation file. If True, it will return the object returned if that
+        file were read with `rdann`. Must not be True if `info_only` is True.
+    verbose : bool, optional
+        Whether to print all the information read about the file (True) or
+        not (False).
+
+    Returns
+    -------
+    N/A : dict, Annotation, optional
+        If 'info_only' is set to True, return all of the annotation
+        information needed to generate WFDB-formatted annotation files.
+        If 'record_only' is set to True, return the WFDB-formatted annotation
+        object generated by the `rdann` output. If none are set to True, write
+        the WFDB-formatted annotation file.
+
+    Notes
+    -----
+    The entire file is composed of (seen here:
+    https://www.edfplus.info/specs/edfplus.html#edfplusannotations):
+
+    HEADER RECORD (we suggest to also adopt the 12 simple additional EDF+ specs)
+    8 ascii : version of this data format (0)
+    80 ascii : local patient identification (mind item 3 of the additional EDF+ specs)
+    80 ascii : local recording identification (mind item 4 of the additional EDF+ specs)
+    8 ascii : startdate of recording (dd.mm.yy) (mind item 2 of the additional EDF+ specs)
+    8 ascii : starttime of recording (hh.mm.ss)
+    8 ascii : number of bytes in header record
+    44 ascii : reserved
+    8 ascii : number of data records (-1 if unknown, obey item 10 of the additional EDF+ specs)
+    8 ascii : duration of a data record, in seconds
+    4 ascii : number of signals (ns) in data record
+    ns * 16 ascii : ns * label (must be 'EDF Annotations')
+    ns * 80 ascii : ns * transducer type (must be whitespace)
+    ns * 8 ascii : ns * physical dimension (must be whitespace)
+    ns * 8 ascii : ns * physical minimum (e.g. -500 or 34, different than physical maximum)
+    ns * 8 ascii : ns * physical maximum (e.g. 500 or 40, different than physical minimum)
+    ns * 8 ascii : ns * digital minimum (must be -32768)
+    ns * 8 ascii : ns * digital maximum (must be 32767)
+    ns * 80 ascii : ns * prefiltering (must be whitespace)
+    ns * 8 ascii : ns * nr of samples in each data record
+    ns * 32 ascii : ns * reserved
+
+    ANNOTATION RECORD
+
+    Examples
+    --------
+    >>> ann_info = wfdb.rdedfann('sample-data/test_edfann.edf')
+
+    """
+    # Some preliminary checks
+    if info_only and record_only:
+        raise Exception(
+            "Both `info_only` and `record_only` are set. Only one "
+            "can be set at a time."
+        )
+
+    # According to the EDF+ docs:
+    #   "The coding is EDF compatible in the sense that old EDF software would
+    #    simply treat this 'EDF Annotations' signal as if it were a (strange-
+    #    looking) ordinary signal"
+    rec = read_edf(
+        record_name,
+        pn_dir=pn_dir,
+        delete_file=delete_file,
+        record_only=True,
+        rdedfann_flag=True,
+    )
+
+    # Convert from array of integers to ASCII strings
+    annotation_string = ""
+    for chunk in rec.p_signal.flatten().astype(np.int64):
+        if chunk + 1 == 0:
+            continue
+        else:
+            adjusted_hex = hex(
+                struct.unpack("<H", struct.pack(">H", chunk + 1))[0]
+            )
+            annotation_string += bytes.fromhex(adjusted_hex[2:]).decode("ascii")
+            # Remove all of the whitespace
+            for rep in ["\x00", "\x14", "\x15"]:
+                annotation_string = annotation_string.replace(rep, " ")
+
+    # Parse the resulting annotation string
+    onsets = []
+    onset_times = []
+    sample_nums = []
+    comments = []
+    durations = []
+    all_anns = annotation_string.split("+")
+    for ann in all_anns:
+        if ann == "":
+            continue
+        try:
+            ann_split = ann.strip().split(" ")
+            onset = float(ann_split[0])
+            hours, rem = divmod(onset, 3600)
+            minutes, seconds = divmod(rem, 60)
+            onset_time = f"{hours:02.0f}:{minutes:02.0f}:{seconds:06.3f}"
+            sample_num = int(onset * rec.sig_len)
+            duration = float(ann_split[1])
+            comment = " ".join(ann_split[2:])
+            if verbose:
+                print(
+                    f"{onset_time}\t{sample_num}\t{comment}\t\tduration: {duration}"
+                )
+            onsets.append(onset)
+            onset_times.append(onset_time)
+            sample_nums.append(sample_num)
+            comments.append(comment)
+            durations.append(duration)
+        except IndexError:
+            continue
+
+    if info_only:
+        return {
+            "onset_time": onset_times,
+            "sample_num": sample_nums,
+            "comment": comments,
+            "duration": durations,
+        }
+    else:
+        df_in = pd.DataFrame(
+            data={
+                "onset": onsets,
+                "duration": durations,
+                "description": comments,
+            }
+        )
+        df_out = format_ann_from_df(df_in)
+        # Remove extension from input file name
+        record_name = record_name.split(os.sep)[-1].split(".")[0]
+        extension = "atr"
+        fs = rec.fs
+        sample = (df_out["onset"].to_numpy() * fs).astype(np.int64)
+        # Assume each annotation is a comment
+        symbol = ['"'] * len(df_out.index)
+        subtype = np.array([22] * len(df_out.index))
+        # Assume each annotation belongs with the 1st channel
+        chan = np.array([0] * len(df_out.index))
+        num = np.array([0] * len(df_out.index))
+        aux_note = df_out["description"].tolist()
+
+        if record_only:
+            return Annotation(
+                record_name=record_name,
+                extension=extension,
+                sample=sample,
+                symbol=symbol,
+                subtype=subtype,
+                chan=chan,
+                num=num,
+                aux_note=aux_note,
+                fs=fs,
+            )
+        else:
+            wrann(
+                record_name,
+                extension,
+                sample=sample,
+                symbol=symbol,
+                subtype=subtype,
+                chan=chan,
+                num=num,
+                aux_note=aux_note,
+                fs=fs,
             )
