+++ wfdb/processing/__init__.py
@@ -11,7 +11,7 @@ from wfdb.processing.evaluate import (
     compare_annotations,
     benchmark_mitdb,
 )
-from wfdb.processing.hr import compute_hr, calc_rr, calc_mean_hr
+from wfdb.processing.hr import compute_hr, calc_rr, calc_mean_hr, ann2rr, rr2ann
 from wfdb.processing.peaks import find_peaks, find_local_peaks, correct_peaks
 from wfdb.processing.qrs import XQRS, xqrs_detect, gqrs_detect
 from wfdb.processing.filter import sigavg
