+++ wfdb/processing/qrs.py
@@ -1,5 +1,4 @@
 import copy
-import pdb
 
 import numpy as np
 from scipy import signal
